arduino/requirements.txt
flask
flask-socketio
pyserial
matplotlib
numpy
scipy